+++ lib/src/models/queue_type.dart
@@ -1,0 +1,5 @@
+enum QueueType {
+  none,
+  perDevice,
+  global,
+}