+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 0.9.12
-* Support all DevicePairingKinds on Windows
+* Support all DevicePairingKinds on Windows 11
+* Add .perDevice queue
+* Improve code level documentation
 
 ## 0.9.11
 * Add device name prefix filtering